another try/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-JWT-Extended
psycopg2-binary
python-dotenv
requests
